+++ daigou-service/integration-test/src/test/java/jiaoni/daigou/service/integrationtest/PingIntegrationTest.java
