+++ daigou-service/integration-test/src/test/resources/customer_only.jpg
