+++ daigou-service/service/src/main/java/jiaoni/daigou/service/appengine/tasks/SyncJiaoniCustomersTaskRunner.java
@@ -9,9 +9,9 @@ import jiaoni.common.location.CnZone;
 import jiaoni.daigou.lib.teddy.TeddyAdmins;
 import jiaoni.daigou.lib.teddy.TeddyClient;
 import jiaoni.daigou.lib.teddy.model.Receiver;
-import jiaoni.daigou.wiremodel.entity.Address;
+import jiaoni.common.wiremodel.Address;
 import jiaoni.daigou.wiremodel.entity.Customer;
-import jiaoni.daigou.wiremodel.entity.PhoneNumber;
+import jiaoni.common.wiremodel.PhoneNumber;
 import jiaoni.daigou.service.appengine.AppEnvs;
 import jiaoni.daigou.service.appengine.impls.CustomerDbClient;
 import org.apache.commons.lang3.StringUtils;
