+++ daigou-service/service/src/main/java/jiaoni/daigou/service/appengine/tasks/SyncJiaoniShippingOrdersTaskRunner.java
@@ -13,7 +13,7 @@ import jiaoni.daigou.lib.teddy.TeddyAdmins;
 import jiaoni.daigou.lib.teddy.TeddyClient;
 import jiaoni.daigou.service.appengine.utils.ShippingOrderUtils;
 import jiaoni.daigou.service.appengine.utils.TeddyUtils;
-import jiaoni.daigou.wiremodel.entity.Price;
+import jiaoni.common.wiremodel.Price;
 import jiaoni.daigou.wiremodel.entity.ShippingOrder;
 import jiaoni.daigou.wiremodel.entity.ShippingOrder.Status;
 import jiaoni.daigou.service.appengine.AppEnvs;
