+++ daigou-service/service/src/main/java/jiaoni/daigou/service/appengine/utils/TeddyUtils.java
@@ -5,11 +5,11 @@ import jiaoni.daigou.lib.teddy.model.Order;
 import jiaoni.daigou.lib.teddy.model.OrderPreview;
 import jiaoni.daigou.lib.teddy.model.Receiver;
 import jiaoni.daigou.lib.teddy.model.ShippingHistoryEntry;
-import jiaoni.daigou.wiremodel.entity.Address;
+import jiaoni.common.wiremodel.Address;
 import jiaoni.daigou.wiremodel.entity.Customer;
-import jiaoni.daigou.wiremodel.entity.PhoneNumber;
+import jiaoni.common.wiremodel.PhoneNumber;
 import jiaoni.daigou.wiremodel.entity.Postman;
-import jiaoni.daigou.wiremodel.entity.Price;
+import jiaoni.common.wiremodel.Price;
 import jiaoni.daigou.wiremodel.entity.Product;
 import jiaoni.daigou.wiremodel.entity.ProductCategory;
 import jiaoni.daigou.wiremodel.entity.ShippingOrder;
