+++ daigou-service/tools/src/main/java/jiaoni/daigou/tools/VerifyDbClient.java
@@ -10,9 +10,9 @@ import jiaoni.common.appengine.access.db.DbClient;
 import jiaoni.common.model.Env;
 import jiaoni.daigou.service.appengine.impls.CustomerDbClient;
 import jiaoni.daigou.tools.remote.RemoteApi;
-import jiaoni.daigou.wiremodel.entity.Address;
+import jiaoni.common.wiremodel.Address;
 import jiaoni.daigou.wiremodel.entity.Customer;
-import jiaoni.daigou.wiremodel.entity.PhoneNumber;
+import jiaoni.common.wiremodel.PhoneNumber;
 import org.apache.commons.lang3.builder.EqualsBuilder;
 import org.apache.commons.lang3.builder.HashCodeBuilder;
 import org.apache.commons.lang3.builder.ToStringBuilder;
