+++ daigou-service/tools/src/main/java/songfan/tools/SendEmail2.java
@@ -18,8 +18,8 @@ import jiaoni.daigou.wiremodel.entity.Customer;
 import jiaoni.daigou.wiremodel.entity.Delivery;
 import jiaoni.daigou.wiremodel.entity.Order;
 import jiaoni.daigou.wiremodel.entity.OrderEntry;
-import jiaoni.daigou.wiremodel.entity.PhoneNumber;
-import jiaoni.daigou.wiremodel.entity.Price;
+import jiaoni.common.wiremodel.PhoneNumber;
+import jiaoni.common.wiremodel.Price;
 import jiaoni.daigou.wiremodel.entity.Product;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.tuple.Pair;
