+++ wire-model/src/main/proto/common/media_object.proto
@@ -1,7 +1,7 @@
 syntax = "proto3";
-package jiaoni.daigou.wiremodel.entity;
+package jiaoni.common.wiremodel;
 
-option java_package = "jiaoni.daigou.wiremodel.entity";
+option java_package = "jiaoni.common.wiremodel";
 option java_multiple_files = true;
 
 message MediaObject {
