+++ wire-model/src/main/proto/daigou/entity/customer.proto
@@ -4,9 +4,7 @@ package jiaoni.daigou.wiremodel.entity;
 option java_package = "jiaoni.daigou.wiremodel.entity";
 option java_multiple_files = true;
 
-import "common/common.proto"
-import "daigou/entity/common.proto";
-import "daigou/entity/media_object.proto";
+import "common/common.proto";
 
 message Customer {
     message SocialContacts {
@@ -16,8 +14,8 @@ message Customer {
 
     string id = 1;
     string name = 2;
-    PhoneNumber phone = 3;
-    repeated Address addresses = 4;
+    jiaoni.common.wiremodel.PhoneNumber phone = 3;
+    repeated jiaoni.common.wiremodel.Address addresses = 4;
     string id_card = 5;
     repeated string emails = 8;
     int32 default_address_index = 9;
