+++ wire-model/src/main/proto/daigou/entity/delivery.proto
@@ -4,11 +4,12 @@ package jiaoni.daigou.wiremodel.entity;
 option java_package = "jiaoni.daigou.wiremodel.entity";
 option java_multiple_files = true;
 
-import "daigou/entity/common.proto";
+import "common/common.proto";
+import "daigou/entity/postman.proto";
 
 message Delivery {
     string delivery_address_raw = 10;
-    Address delivery_address = 11;
+    jiaoni.common.wiremodel.Address delivery_address = 11;
 
     string delivery_time_raw = 20;
     int64 delivery_time = 21;
