+++ wire-model/src/main/proto/daigou/entity/postman.proto
@@ -4,7 +4,9 @@ package jiaoni.daigou.wiremodel.entity;
 option java_package = "jiaoni.daigou.wiremodel.entity";
 option java_multiple_files = true;
 
+import "common/common.proto";
+
 message Postman {
     string name = 2;
-    PhoneNumber phone = 3;
+    jiaoni.common.wiremodel.PhoneNumber phone = 3;
 }
