+++ wire-model/src/main/proto/daigou/interface/parse_interface_model.proto
@@ -4,7 +4,7 @@ package jiaoni.daigou.wiremodel.api;
 option java_package = "jiaoni.daigou.wiremodel.api";
 option java_multiple_files = true;
 
-import "daigou/entity/common.proto";
+import "common/common.proto";
 import "daigou/entity/customer.proto";
 import "daigou/entity/product.proto";
 
@@ -39,7 +39,7 @@ message ParsedObject {
     int32 confidence = 1;
 
     oneof content {
-        jiaoni.daigou.wiremodel.entity.Address address = 50;
+        jiaoni.common.wiremodel.Address address = 50;
         jiaoni.daigou.wiremodel.entity.Customer customer = 51;
         jiaoni.daigou.wiremodel.entity.Product product = 52;
     }
