+++ wire-model/src/main/proto/daigou/interface/shopping_list_interface_model.proto
@@ -4,7 +4,7 @@ package jiaoni.daigou.wiremodel.api;
 option java_package = "jiaoni.daigou.wiremodel.api";
 option java_multiple_files = true;
 
-import "daigou/entity/common.proto";
+import "common/common.proto";
 import "daigou/entity/shopping_list_item.proto";
 
 message InitShoppingListItemRequest {
@@ -20,7 +20,7 @@ message AssignOwnershipShoppingListItemRequest {
 
 message PurchaseShoppingListItemRequest {
     string purchaser_name = 1;
-    jiaoni.daigou.wiremodel.entity.Price total_purchase_price = 2;
+    jiaoni.common.wiremodel.Price total_purchase_price = 2;
     string purchasing_source = 3;
     repeated string media_ids = 4;
 }
