+++ wire-model/src/main/proto/songfan/entity/menu.proto
@@ -1,0 +1,27 @@
+syntax = "proto3";
+package jiaoni.songfan.wiremodel.entity;
+
+option java_package = "jiaoni.songfan.wiremodel.entity";
+option java_multiple_files = true;
+
+import "common/common.proto";
+import "songfan/entity/combo.proto";
+import "songfan/entity/dish.proto";
+
+message Menu {
+    string id = 1;
+    string name = 2;
+    int64 creation_time = 3;
+    int64 expiration_time = 4;
+
+    message MenuEntry {
+        oneof entry {
+            Combo combo = 1;
+            Dish dish =2;
+        }
+        jiaoni.common.wiremodel.Price price = 10;
+    }
+    repeated MenuEntry menu_entries = 10;
+}
+
+
